+++ examples/brawlapi/async.py
@@ -1,7 +1,7 @@
 import brawlstats
 import asyncio
 
-client = brawlstats.Client('token', is_async=True)
+client = brawlstats.BrawlAPI('token', is_async=True)
 # Do not post your token on a public github
 
 # await only works in an async loop
@@ -12,25 +12,25 @@ async def main():
 
     club = await player.get_club()
     print(club.tag)
-    best_players = club.members[0:3]  # members sorted by trophies, gets best 3 players
+    best_players = club.members[:5]  # members sorted by trophies, gets best 5 players
     for player in best_players:
         print(player.name, player.trophies)
 
-    leaderboard = await client.get_leaderboard('players', count=5)  # gets top 5 players
+    leaderboard = await client.get_leaderboard('players', limit=5)  # gets top 5 players
     for player in leaderboard:
         print(player.name, player.position)
 
     events = await client.get_events()
     print(events.current[0].map_name)
 
+    battles = await client.get_battle_logs('GGJVJLU2')
+    print(battles[0].battle.mode)
+
     misc = await client.get_misc()
     print(misc.time_until_season_ends)
 
     search = await client.search_club('Cactus Bandits')
     print(search[0].tag)
-
-    battles = await client.get_battle_logs('GGJVJLU2')
-    print(battles[0].battle.mode)
 
 # run the async loop
 loop = asyncio.get_event_loop()