+++ tests/brawlapi/test_blocking.py
@@ -1,5 +1,4 @@
 import datetime
-import logging
 import unittest
 import os
 import time
@@ -7,26 +6,24 @@ import time
 import brawlstats
 from dotenv import load_dotenv, find_dotenv
 
-load_dotenv(find_dotenv('./.env'))
+load_dotenv(find_dotenv('../.env'))
 
-TOKEN = os.getenv('token')
+TOKEN = os.getenv('unofficial_token')
 
 
 class TestBlockingClient(unittest.TestCase):
-    """Tests all methods in the blocking client that
+    """Tests all methods in the blocking BrawlAPI client that
     uses the `requests` module in `brawlstats`
     """
 
     def setUp(self):
         self.player_tag = 'GGJVJLU2'
         self.club_tag = 'QCGV8PG'
-        self.client = brawlstats.Client(
+        self.client = brawlstats.BrawlAPI(
             TOKEN,
             is_async=False,
-            timeout=30,
-            debug=True
+            timeout=30
         )
-        logging.basicConfig(level=logging.DEBUG)
 
     def tearDown(self):
         time.sleep(1)
@@ -42,17 +39,17 @@ class TestBlockingClient(unittest.TestCase):
 
     def test_get_leaderboard_player(self):
         lb = self.client.get_leaderboard('players')
-        self.assertTrue(isinstance(lb, brawlstats.Leaderboard))
+        self.assertTrue(isinstance(lb, brawlstats.brawlapi.Leaderboard))
         region = self.client.get_leaderboard('players', region='us')
-        self.assertTrue(isinstance(region, brawlstats.Leaderboard))
+        self.assertTrue(isinstance(region, brawlstats.brawlapi.Leaderboard))
 
     def test_get_leaderboard_club(self):
         lb = self.client.get_leaderboard('clubs')
-        self.assertTrue(isinstance(lb, brawlstats.Leaderboard))
+        self.assertTrue(isinstance(lb, brawlstats.brawlapi.Leaderboard))
 
     def test_get_leaderboard_brawler(self):
-        lb = self.client.get_leaderboard('shelly')
-        self.assertTrue(isinstance(lb, brawlstats.Leaderboard))
+        lb = self.client.get_leaderboard('brawlers', brawler='shelly')
+        self.assertTrue(isinstance(lb, brawlstats.brawlapi.Leaderboard))
 
     def test_get_events(self):
         events = self.client.get_events()
@@ -60,9 +57,9 @@ class TestBlockingClient(unittest.TestCase):
 
     def test_get_constants(self):
         default = self.client.get_constants()
-        self.assertTrue(isinstance(default, brawlstats.Constants))
+        self.assertTrue(isinstance(default, brawlstats.brawlapi.Constants))
         maps = self.client.get_constants('maps')
-        self.assertTrue(isinstance(maps, brawlstats.Constants))
+        self.assertTrue(isinstance(maps, brawlstats.brawlapi.Constants))
         get_constants = self.client.get_constants
         invalid_key = 'invalid'
         self.assertRaises(KeyError, get_constants, invalid_key)
@@ -77,32 +74,32 @@ class TestBlockingClient(unittest.TestCase):
 
     def test_battle_logs(self):
         logs = self.client.get_battle_logs(self.player_tag)
-        self.assertTrue(isinstance(logs, list))
+        self.assertTrue(isinstance(logs, brawlstats.brawlapi.BattleLog))
 
     # Other
     def test_invalid_tag(self):
-        get_profile = self.client.get_profile
+        get_player = self.client.get_player
         invalid_tag = 'P'
-        self.assertRaises(brawlstats.NotFoundError, get_profile, invalid_tag)
+        self.assertRaises(brawlstats.NotFoundError, get_player, invalid_tag)
         invalid_tag = 'AAA'
-        self.assertRaises(brawlstats.NotFoundError, get_profile, invalid_tag)
+        self.assertRaises(brawlstats.NotFoundError, get_player, invalid_tag)
         invalid_tag = '2PPPPPPP'
-        self.assertRaises(brawlstats.NotFoundError, get_profile, invalid_tag)
+        self.assertRaises(brawlstats.NotFoundError, get_player, invalid_tag)
 
     def test_invalid_lb(self):
         get_lb = self.client.get_leaderboard
         invalid_type = 'test'
-        invalid_count = 200
-        self.assertRaises(ValueError, get_lb, invalid_type, invalid_count)
+        invalid_limit = 200
+        self.assertRaises(ValueError, get_lb, invalid_type, invalid_limit)
         invalid_type = 'players'
-        invalid_count = 'string'
-        self.assertRaises(ValueError, get_lb, invalid_type, invalid_count)
+        invalid_limit = 'string'
+        self.assertRaises(ValueError, get_lb, invalid_type, invalid_limit)
         invalid_type = 'players'
-        invalid_count = 201
-        self.assertRaises(ValueError, get_lb, invalid_type, invalid_count)
+        invalid_limit = 201
+        self.assertRaises(ValueError, get_lb, invalid_type, invalid_limit)
         invalid_type = 'players'
-        invalid_count = -5
-        self.assertRaises(ValueError, get_lb, invalid_type, invalid_count)
+        invalid_limit = -5
+        self.assertRaises(ValueError, get_lb, invalid_type, invalid_limit)
 
 
 if __name__ == '__main__':